Start to detect hashtag patterns in text
